requirements.txt
absl-py
pytest
pytest-xdist
isort
black
flake8
flake8-blind-except
flake8-builtins
flake8-debugger
flake8-isort
gym
imitation @ git+https://github.com/HumanCompatibleAI/imitation.git
matplotlib
mujoco_py
numpy
pandas
seaborn
setuptools
scipy
stable-baselines @ git+https://github.com/hill-a/stable-baselines.git
tensorflow>=1.13,<1.14
xarray
